Use NameData and namestrcpy for names
Using `strlcpy` to copy variables holding PostgreSQL names can cause
issues since names are fixed-size types of length 64. This means that
any data that follows the initial null-terminated string will also be
part of the data.

Instead of using `const char*` for PostgreSQL names, use `NameData`
type for PostgreSQL names and use `namestrcpy` to copy them rather
than `strlcpy`.

(cherry picked from commit 38b71d0e703caa568a38a854e041a916c5601057)

+++ scripts/test_update_from_tag.sh
@@ -43,8 +43,12 @@ remove_containers() {
     docker rm -vf ${CONTAINER_CLEAN_RESTORE} 2>/dev/null
     docker rm -vf ${CONTAINER_UPDATED}  2>/dev/null
     docker rm -vf ${CONTAINER_CLEAN_RERUN} 2>/dev/null
-    docker volume rm -f ${CLEAN_VOLUME} 2>/dev/null
-    docker volume rm -f ${UPDATE_VOLUME} 2>/dev/null
+    if [[ -n "${CLEAN_VOLUME}" ]]; then
+	docker volume rm -f ${CLEAN_VOLUME} 2>/dev/null
+    fi
+    if [[ -n "${UPDATE_VOLUME}" ]]; then
+	docker volume rm -f ${UPDATE_VOLUME} 2>/dev/null
+    fi
 }
 
 cleanup() {
